+++ babel.config.cjs
@@ -7,18 +7,17 @@ module.exports = {
                 corejs: '3.32',
             },
         ],
-        'solid',
+        [
+            '@babel/preset-react',
+            {
+                runtime: 'automatic',
+            },
+        ],
         '@babel/preset-typescript',
     ],
     plugins: [
         '@babel/plugin-syntax-dynamic-import',
         '@babel/plugin-proposal-class-properties',
         '@babel/plugin-proposal-object-rest-spread',
-        [
-            'solid-refresh/babel',
-            {
-                bundler: 'webpack5', // or "rspack-esm"
-            },
-        ],
     ],
 };